requirements.txt
django==1.10.4
djangorestframework==3.5.1
django-rest-swagger==2.1.0
django-cors-headers
